+++ src/rgb/formats/bgr_888.rs
@@ -11,6 +11,7 @@ use crate::rgb::Bgr;
 ///   uint8_t b;
 ///   uint8_t g;
 ///   uint8_t r;
+///   // Padding: 8 bits
 /// }
 /// ```
 ///