+++ src/rgb/macros/impl_with_alpha_packed.rs
@@ -1,11 +1,13 @@
 macro_rules! impl_with_alpha_packed {
     ($ty:ident, $alpha_shift:expr, $alpha_mask:expr, $alpha_clear:expr) => {
-        impl crate::alpha::HasAlpha<u8> for $ty {
-            fn alpha(&self) -> u8 {
-                ((self.packed >> $alpha_shift) & $alpha_mask) as u8
+        impl crate::alpha::HasAlpha for $ty {
+            type Component = u8;
+
+            fn alpha(&self) -> Self::Component {
+                ((self.packed >> $alpha_shift) & $alpha_mask) as Self::Component
             }
 
-            fn set_alpha(&mut self, value: u8) {
+            fn set_alpha(&mut self, value: Self::Component) {
                 self.packed = (self.packed & $alpha_clear)
                     | ((u16::from(value) & $alpha_mask) << $alpha_shift);
             }