+++ src/rgb/traits/rgb_color.rs
@@ -1,10 +1,24 @@
 use crate::rgb::{HasBlue, HasGreen, HasRed};
 
 /// A trait for types that have red, green, and blue components.
-pub trait RgbColor<T>: Sized + Default + HasRed<T> + HasGreen<T> + HasBlue<T> {
+pub trait RgbColor: Sized + Default + HasRed + HasGreen + HasBlue {
     /// Creates a new color with the given red, green, and blue components.
+    ///
+    /// How components are represented, including possible clamping or conversion, is determined by
+    /// the implementation.
+    ///
+    /// This method is provided as a convenience to create _any_ [`RgbColor`] type; most types will
+    /// have their own dedicated constructor methods that may be more efficient and specific. For
+    /// example [`Rgb888::from_rgb`][], [`Rgb565::from_rgb`][], etc.
+    ///
+    /// [`Rgb888::from_rgb`]: crate::rgb::Rgb888::from_rgb
+    /// [`Rgb565::from_rgb`]: crate::rgb::Rgb565::from_rgb
     #[must_use]
-    fn from_rgb(red: T, green: T, blue: T) -> Self {
+    fn from_rgb(
+        red: <Self as crate::rgb::HasRed>::Component,
+        green: <Self as crate::rgb::HasGreen>::Component,
+        blue: <Self as crate::rgb::HasBlue>::Component,
+    ) -> Self {
         let mut color = Self::default();
         color.set_red(red);
         color.set_green(green);
@@ -14,12 +28,18 @@ pub trait RgbColor<T>: Sized + Default + HasRed<T> + HasGreen<T> + HasBlue<T> {
 
     /// Returns the inner representation of the color as a tuple of red, green, and blue components.
     #[must_use]
-    fn into_rgb(self) -> (T, T, T) {
+    fn into_rgb(
+        self,
+    ) -> (
+        <Self as crate::rgb::HasRed>::Component,
+        <Self as crate::rgb::HasGreen>::Component,
+        <Self as crate::rgb::HasBlue>::Component,
+    ) {
         (self.red(), self.green(), self.blue())
     }
 }
 
-impl<T, C> RgbColor<C> for T where T: HasRed<C> + HasGreen<C> + HasBlue<C> + Default + Sized {}
+impl<T> RgbColor for T where T: HasRed + HasGreen + HasBlue + Default + Sized {}
 
 #[cfg(test)]
 mod tests {
@@ -29,7 +49,7 @@ mod tests {
     fn rgb_color_trait() {
         use crate::rgb::{Rgb888, RgbColor};
 
-        let color: Rgb888 = RgbColor::<u8>::from_rgb(255, 0, 0);
+        let color: Rgb888 = RgbColor::from_rgb(255, 0, 0);
         assert_eq!(color.red(), 255);
         assert_eq!(color.green(), 0);
         assert_eq!(color.blue(), 0);