+++ src/rgb/traits/rgba_color.rs
@@ -4,12 +4,25 @@ use crate::{
 };
 
 /// A trait for types that have red, green, blue, and alpha components.
-pub trait RgbaColor<T>:
-    Sized + Default + HasRed<T> + HasGreen<T> + HasBlue<T> + HasAlpha<T>
-{
+pub trait RgbaColor: Sized + Default + HasRed + HasGreen + HasBlue + HasAlpha {
     /// Creates a new color with the given red, green, blue, and alpha components.
+    ///
+    /// How components are represented, including possible clamping or conversion, is determined by
+    /// the implementation.
+    ///
+    /// This method is provided as a convenience to create _any_ [`RgbaColor`] type; most types will
+    /// have their own dedicated constructor methods that may be more efficient and specific. For
+    /// example [`Abgr8888::from_rgba`][], [`Rgbaf32::from_rgba`][], etc.
+    ///
+    /// [`Abgr8888::from_rgba`]: crate::rgb::Abgr8888::from_rgba
+    /// [`Rgbaf32::from_rgba`]: crate::rgb::Rgbaf32::from_rgba
     #[must_use]
-    fn from_rgba(red: T, green: T, blue: T, alpha: T) -> Self {
+    fn from_rgba(
+        red: <Self as crate::rgb::HasRed>::Component,
+        green: <Self as crate::rgb::HasGreen>::Component,
+        blue: <Self as crate::rgb::HasBlue>::Component,
+        alpha: <Self as crate::alpha::HasAlpha>::Component,
+    ) -> Self {
         let mut color = Self::default();
         color.set_red(red);
         color.set_green(green);
@@ -20,12 +33,19 @@ pub trait RgbaColor<T>:
 
     /// Returns the inner representation of the color as a tuple of red, green, blue, and alpha components.
     #[must_use]
-    fn into_rgba(self) -> (T, T, T, T) {
+    fn into_rgba(
+        self,
+    ) -> (
+        <Self as crate::rgb::HasRed>::Component,
+        <Self as crate::rgb::HasGreen>::Component,
+        <Self as crate::rgb::HasBlue>::Component,
+        <Self as crate::alpha::HasAlpha>::Component,
+    ) {
         (self.red(), self.green(), self.blue(), self.alpha())
     }
 }
 
-impl<T, C> RgbaColor<C> for T where T: HasRed<C> + HasGreen<C> + HasBlue<C> + HasAlpha<C> + Default {}
+impl<T> RgbaColor for T where T: HasRed + HasGreen + HasBlue + HasAlpha + Default {}
 
 #[cfg(test)]
 mod tests {
@@ -33,7 +53,7 @@ mod tests {
 
     #[test]
     fn rgba_color_trait() {
-        let color: Abgr8888 = RgbaColor::<u8>::from_rgba(255, 0, 0, 255);
+        let color: Abgr8888 = RgbaColor::from_rgba(255, 0, 0, 255);
         assert_eq!(color.red(), 255);
         assert_eq!(color.green(), 0);
         assert_eq!(color.blue(), 0);